+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.47.0
 
+* Adding progress tracking for Cruise Control rebalances
 * Add support for Kafka 3.9.1
 * Fixed MirrorMaker 2 client rack init container override being ignored.
 * Support for Kubernetes Image Volumes to mount custom plugins